+++ base/sysimg.jl
@@ -51,7 +51,7 @@ convert(::Type{T}, arg)  where {T<:VecElement} = T(arg)
 convert(::Type{T}, arg::T) where {T<:VecElement} = arg
 
 # init core docsystem
-import Core: @doc, @__doc__, @doc_str
+import Core: @doc, @__doc__, @doc_str, WrappedException
 if isdefined(Core, :Inference)
     import Core.Inference.CoreDocs
     Core.atdoc!(CoreDocs.docm)
@@ -72,7 +72,8 @@ end
 ## Load essential files and libraries
 include("essentials.jl")
 include("ctypes.jl")
-include("base.jl")
+include("gcutils.jl")
+include("nullabletype.jl")
 include("generator.jl")
 include("reflection.jl")
 include("options.jl")
@@ -162,6 +163,12 @@ include("reshapedarray.jl")
 include("bitarray.jl")
 include("intset.jl")
 include("associative.jl")
+
+if !isdefined(Core, :Inference)
+    include("docs/core.jl")
+    Core.atdoc!(CoreDocs.docm)
+end
+
 include("dict.jl")
 include("set.jl")
 include("iterators.jl")
@@ -187,11 +194,6 @@ include(string((length(Core.ARGS)>=2 ? Core.ARGS[2] : ""), "version_git.jl")) # 
 
 include("osutils.jl")
 include("c.jl")
-
-if !isdefined(Core, :Inference)
-    include("docs/core.jl")
-    Core.atdoc!(CoreDocs.docm)
-end
 
 # Core I/O
 include("io.jl")
@@ -323,8 +325,8 @@ include("hashing2.jl")
 include("irrationals.jl")
 
 # random number generation
-include("dSFMT.jl")
-include("random.jl")
+include("random/dSFMT.jl")
+include("random/random.jl")
 importall .Random
 
 # (s)printf macros
@@ -416,7 +418,6 @@ include("threadcall.jl")
 include("deprecated.jl")
 
 # Some basic documentation
-include("docs/helpdb.jl")
 include("docs/basedocs.jl")
 
 # Documentation -- should always be included last in sysimg.